requirements.txt
fastapi==0.103.0
numpy==1.24.3
Pillow==10.0.0
python-dotenv==1.0.0
python-multipart==0.0.6
tensorflow==2.13.0
uvicorn==0.23.2
opencv-python==4.8.1.78
face-recognition==1.3.0
